+++ CHANGELOG.rst
@@ -1,6 +1,6 @@
 **3.1.3 - 12/02/24**
 
- - Provide access to relative risk pipeline. 
+ - Feature: Create relative risk pipeline
 
 **3.1.2 - 11/13/24**